Fix issue with shorthand cell type names causing missing descriptions and images, e.g. T for T cell
